requirements.txt
python-dotenv>=1.0.0
google-cloud-storage>=2.0.0
Pillow>=10.0.0
mss>=9.0.0
networkx>=3.0.0
google-generativeai>=0.3.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
